+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [1.1.0](https://github.com/pascaliske/eslint-config/compare/v1.0.0...v1.1.0) (2020-10-11)
+
+
+### Features
+
+* **rules:** implement eslint rules ([114485f](https://github.com/pascaliske/eslint-config/commit/114485f51bbed50911abd44698696cb5bfb5dfec))
+* **rules:** implement typescript specific rules ([eec05ac](https://github.com/pascaliske/eslint-config/commit/eec05ac30b9a1935e2db4a66b89722715ad5579e))
+
+
+
 # 1.0.0 (2020-06-09)